placing spm test file in the right place
